Add git-gone(1) manpage
Add new manpage for git-gone to make "git gone --help" work.  The man
page builds from Asciidoc source via asciidoctor.  To avoid duplication
remove README.md and symlink README.adoc to the manpage source.

We commit the actual manpage generated from the manpage source as well
to avoid a build dependency on asciidoctor.

While making a release update the version number and revision date in
the man page source and rebuild the manpage to include the change in the
release commit.

Add new CI job to build the manpage and check whether the generated
manpage is up to date.